Refactor tests: split character, quantifier, and anchor tests
Moved character class, quantifier, and anchor tests from Rejigs.Tests.cs into separate files: Characters.cs, Quantifiers.cs, and Anchors.cs. This improves test organization and maintainability.
