+++ Rejigs.Tests/Rejigs.Tests.cs
@@ -5,249 +5,7 @@ namespace Rejigs.Tests
 {
     public class RejigsTests
     {
-        #region Character Class Tests
-
-        [Test]
-        public void AnyDigit_MatchesSingleDigit()
-        {
-            var regex = Rejigs.Create().AnyDigit().Build();
-            Assert.That("5", Does.Match(regex));
-            Assert.That("a", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyNonDigit_MatchesSingleNonDigit()
-        {
-            var regex = Rejigs.Create().AnyNonDigit().Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That("5", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyLetterOrDigit_MatchesWordCharacter()
-        {
-            var regex = Rejigs.Create().AnyLetterOrDigit().Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That("_", Does.Match(regex));
-            Assert.That("$", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyNonLetterOrDigit_MatchesNonWordCharacter()
-        {
-            var regex = Rejigs.Create().AnyNonLetterOrDigit().Build();
-            Assert.That("$", Does.Match(regex));
-            Assert.That("a", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnySpace_MatchesWhitespace()
-        {
-            var regex = Rejigs.Create().AnySpace().Build();
-            Assert.That(" ", Does.Match(regex));
-            Assert.That("\t", Does.Match(regex));
-            Assert.That("a", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyNonSpace_MatchesNonWhitespace()
-        {
-            var regex = Rejigs.Create().AnyNonSpace().Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That(" ", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyCharacter_MatchesAnyCharacter()
-        {
-            var regex = Rejigs.Create().AnyCharacter().Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That("1", Does.Match(regex));
-            Assert.That("$", Does.Match(regex));
-            Assert.That("\n", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyOf_MatchesCharactersInSet()
-        {
-            var regex = Rejigs.Create().AnyOf("abc").Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That("b", Does.Match(regex));
-            Assert.That("c", Does.Match(regex));
-            Assert.That("z", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyInRange_MatchesCharactersInRange()
-        {
-            var regex = Rejigs.Create().AnyInRange('a', 'f').Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That("d", Does.Match(regex));
-            Assert.That("f", Does.Match(regex));
-            Assert.That("g", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AnyExcept_MatchesCharactersNotInSet()
-        {
-            var regex = Rejigs.Create().AnyExcept("xyz").Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That("x", Does.Not.Match(regex));
-            Assert.That("y", Does.Not.Match(regex));
-        }
-
-        #endregion
-
-        #region Quantifier Tests
-
-        [Test]
-        public void ZeroOrMore_WithText_MatchesOptionalRepeatedText()
-        {
-            var regex = Rejigs.Create()
-                .Text("pre")
-                .ZeroOrMore(r => r.Text("-"))
-                .Text("fix")
-                .Build();
-
-            Assert.That("prefix", Does.Match(regex));
-            Assert.That("pre-fix", Does.Match(regex));
-            Assert.That("pre--fix", Does.Match(regex));
-            Assert.That("pre**fix", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void ZeroOrMore_CombinedWithOtherPatterns()
-        {
-            var regex = Rejigs.Create()
-                .AtStart()
-                .Text("A")
-                .ZeroOrMore(r => r.AnyDigit())
-                .Text("Z")
-                .AtEnd()
-                .Build();
-
-            Assert.That("AZ", Does.Match(regex));
-            Assert.That("A1Z", Does.Match(regex));
-            Assert.That("A123Z", Does.Match(regex));
-            Assert.That("A1B2Z", Does.Not.Match(regex));
-            Assert.That("ABC", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void OneOrMore_MatchesRepeatedPattern()
-        {
-            var regex = Rejigs.Create().OneOrMore(r => r.Text("a")).Build();
-            Assert.That("a", Does.Match(regex));
-            Assert.That("aaa", Does.Match(regex));
-            Assert.That("", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void Optional_MatchesOptionalPattern()
-        {
-            var regex = Rejigs.Create().Optional(r => r.Text("a")).Build();
-            Assert.That("", Does.Match(regex));
-            Assert.That("a", Does.Match(regex));
-            Assert.That("aa", Does.Match(regex));
-        }
-
-        [Test]
-        public void Optional_WithText_MatchesOptionalText()
-        {
-            var regex = Rejigs.Create()
-                .Text("colou")
-                .Optional("r")
-                .Build();
-
-            Assert.That("colo", Does.Not.Match(regex));
-            Assert.That("colou", Does.Match(regex));
-            Assert.That("colour", Does.Match(regex));
-            Assert.That("colours", Does.Match(regex));
-        }
-
-        [Test]
-        public void Optional_WithPattern_MatchesOptionalPattern()
-        {
-            var regex = Rejigs.Create()
-                .Optional(r => r.Text("pre-"))
-                .Text("fix")
-                .Build();
-
-            Assert.That("fix", Does.Match(regex));
-            Assert.That("pre-fix", Does.Match(regex));
-        }
-
-        [Test]
-        public void Exactly_MatchesExactCount()
-        {
-            var regex = Rejigs.Create().AtStart().AnyDigit().Exactly(3).AtEnd().Build();
-            Assert.That("123", Does.Match(regex));
-            Assert.That("12", Does.Not.Match(regex));
-            Assert.That("1234", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AtLeast_MatchesMinimumCount()
-        {
-            var regex = Rejigs.Create().AnyDigit().AtLeast(2).Build();
-            Assert.That("12", Does.Match(regex));
-            Assert.That("123", Does.Match(regex));
-            Assert.That("1", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void Between_MatchesRangeOfCounts()
-        {
-            var regex = Rejigs.Create().AnyDigit().Between(2, 4).Build();
-            Assert.That("12", Does.Match(regex));
-            Assert.That("123", Does.Match(regex));
-            Assert.That("1234", Does.Match(regex));
-            Assert.That("1", Does.Not.Match(regex));
-            Assert.That("12345", Does.Match(regex));
-        }
-
-        #endregion
-
-        #region Anchor Tests
-
-        [Test]
-        public void AtStart_AnchorsBeginingOfLine()
-        {
-            var regex = Rejigs.Create().AtStart().Text("abc").Build();
-            Assert.That("abc", Does.Match(regex));
-            Assert.That("abcdef", Does.Match(regex));
-            Assert.That("xabc", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AtEnd_AnchorsEndOfLine()
-        {
-            var regex = Rejigs.Create().Text("abc").AtEnd().Build();
-            Assert.That("abc", Does.Match(regex));
-            Assert.That("xabc", Does.Match(regex));
-            Assert.That("abcx", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void AtWordBoundary_MatchesAtWordBoundary()
-        {
-            var regex = Rejigs.Create().AtWordBoundary().Text("cat").Build();
-            Assert.That("cat", Does.Match(regex));
-            Assert.That("the cat", Does.Match(regex));
-            Assert.That("tomcat", Does.Not.Match(regex));
-        }
-
-        [Test]
-        public void NotAtWordBoundary_MatchesNotAtWordBoundary()
-        {
-            var regex = Rejigs.Create().NotAtWordBoundary().Text("cat").Build();
-            Assert.That("tomcat", Does.Match(regex));
-            Assert.That("cat", Does.Not.Match(regex));
-        }
-
-        #endregion
-
-        #region Group Tests
+       #region Group Tests
 
         [Test]
         public void Grouping_WithQuantifier()
